MKLDNN plugin extension
